Split app unit tests by domain
