+++ ASMR WalkTests/ASMR_WalkTests.swift
@@ -2,8 +2,6 @@
 //  ASMR_WalkTests.swift
 //  ASMR WalkTests
 //
-//  Created by David Heath on 5/24/26.
-//
 
 import Foundation
 import CoreLocation
@@ -184,1460 +182,5 @@ struct ASMR_WalkTests {
     func videoLifecycleStopsOnSceneDeactivation() {
         #expect(RecordingLifecyclePolicy.shouldStopVideoWalkWhenSceneDeactivates(isRecordingVideoWalk: true))
         #expect(RecordingLifecyclePolicy.shouldStopVideoWalkWhenSceneDeactivates(isRecordingVideoWalk: false) == false)
-    }
-
-    @Test("Privacy usage descriptions are specific")
-    func privacyUsageDescriptionsAreSpecific() throws {
-        let projectRoot = try #require(repositoryRoot())
-        let projectSettingsURL = projectRoot.appending(path: "ASMR Walk.xcodeproj/project.pbxproj")
-        let projectSettingsText = try String(contentsOf: projectSettingsURL, encoding: .utf8)
-
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSLocationWhenInUseUsageDescription = \"ASMR Walk uses your location while recording to draw and save your walking route.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSLocationAlwaysAndWhenInUseUsageDescription = \"ASMR Walk uses background location only when you enable background GPS recording for walks.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSCameraUsageDescription = \"ASMR Walk uses the camera to record video walks.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSMicrophoneUsageDescription = \"ASMR Walk uses the microphone to record video walks.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSPhotoLibraryAddUsageDescription = \"ASMR Walk saves a copy of a video walk to Photos when you choose Save Video to Photos.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_KEY_NSPhotoLibraryUsageDescription = \"ASMR Walk reads older Photos-backed video walks so you can replay them with your route.\";"))
-        #expect(projectSettingsText.contains("INFOPLIST_FILE = \"ASMR-Walk-Info.plist\";"))
-        #expect(projectSettingsText.contains("CODE_SIGN_ENTITLEMENTS = \"ASMR Walk/ASMR Walk.entitlements\";"))
-
-        let entitlementsURL = projectRoot.appending(path: "ASMR Walk/ASMR Walk.entitlements")
-        let entitlementsData = try Data(contentsOf: entitlementsURL)
-        let entitlements = try #require(
-            PropertyListSerialization.propertyList(from: entitlementsData, format: nil) as? [String: Any]
-        )
-        #expect(entitlements["com.apple.developer.icloud-services"] as? [String] == ["CloudKit"])
-        #expect(entitlements["com.apple.developer.icloud-container-identifiers"] as? [String] == [CloudSyncConfiguration.containerIdentifier])
-
-        if let infoPlist = try staticInfoPlist(in: projectRoot) {
-            #expect(infoPlist["UIBackgroundModes"] as? [String] == ["location", "remote-notification"])
-        }
-    }
-
-    private func repositoryRoot() -> URL? {
-        let environment = ProcessInfo.processInfo.environment
-        let candidates = [
-            URL(fileURLWithPath: #filePath),
-            URL(fileURLWithPath: FileManager.default.currentDirectoryPath),
-            Bundle.main.bundleURL,
-            environment["SRCROOT"].map(URL.init(fileURLWithPath:)),
-            environment["SOURCE_ROOT"].map(URL.init(fileURLWithPath:)),
-            environment["PROJECT_DIR"].map(URL.init(fileURLWithPath:)),
-            environment["CI_WORKSPACE"].map { URL(fileURLWithPath: $0).appending(path: "repository") }
-        ].compactMap { $0 }
-
-        for candidate in candidates {
-            if let root = repositoryRoot(startingAt: candidate) {
-                return root
-            }
-        }
-
-        return nil
-    }
-
-    private func repositoryRoot(startingAt url: URL) -> URL? {
-        var candidate = url.hasDirectoryPath ? url : url.deletingLastPathComponent()
-
-        while candidate.path != candidate.deletingLastPathComponent().path {
-            let projectSettingsURL = candidate.appending(path: "ASMR Walk.xcodeproj/project.pbxproj")
-            if FileManager.default.fileExists(atPath: projectSettingsURL.path) {
-                return candidate
-            }
-
-            candidate.deleteLastPathComponent()
-        }
-
-        return nil
-    }
-
-    private func staticInfoPlist(in projectRoot: URL) throws -> [String: Any]? {
-        let candidates = [
-            projectRoot.appending(path: "ASMR-Walk-Info.plist"),
-            projectRoot.appending(path: "ASMR Walk/ASMR-Walk-Info.plist"),
-            projectRoot.appending(path: "ASMR Walk/ASMR Walk/ASMR-Walk-Info.plist")
-        ]
-
-        guard let infoPlistURL = candidates.first(where: { FileManager.default.fileExists(atPath: $0.path) }) else {
-            return nil
-        }
-
-        let infoPlistData = try Data(contentsOf: infoPlistURL)
-        return try PropertyListSerialization.propertyList(
-            from: infoPlistData,
-            format: nil
-        ) as? [String: Any]
-    }
-
-    @Test("Cloud sync uses the configured private container")
-    func cloudSyncConfiguration() {
-        #expect(CloudSyncConfiguration.containerIdentifier == "iCloud.com.bald-traveler.ASMRWalk")
-    }
-
-    @Test("Cloud sync maps iCloud account states to user-visible states")
-    func cloudSyncAccountStateMapping() {
-        #expect(CloudSyncStatus.state(for: .available) == .available)
-        #expect(CloudSyncStatus.state(for: .noAccount) == .noAccount)
-        #expect(CloudSyncStatus.state(for: .restricted) == .restricted)
-        #expect(CloudSyncStatus.state(for: .temporarilyUnavailable) == .temporarilyUnavailable)
-        #expect(CloudSyncStatus.state(for: .couldNotDetermine) == .couldNotDetermine)
-        #expect(CloudSyncAccountState.available.message.contains("Video files stay on the device"))
-    }
-
-    @Test("Delete confirmation explains Photos video ownership")
-    func deleteConfirmationForPhotosVideo() {
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoAssetIdentifier: "photos-asset-id"
-        )
-
-        #expect(recording.deleteConfirmationMessage.contains("video remains in Photos"))
-    }
-
-    @Test("Local-only video messaging explains cross-device availability")
-    func localOnlyVideoMessaging() {
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoURL: URL(filePath: "/tmp/missing-local-video.mov")
-        )
-
-        #expect(recording.videoStorage == .localOnly)
-        #expect(recording.videoAvailabilityTitle == "Video Not on This Device")
-        #expect(recording.videoAvailabilityMessage.contains("video file stays on the device where it was recorded"))
-    }
-
-    @Test("Missing local thumbnails are detectable after sync")
-    func missingLocalThumbnailDetection() {
-        let recording = WalkRecording(
-            title: "Synced Walk",
-            mode: .walk,
-            thumbnailURL: URL(filePath: "/tmp/missing-route-thumbnail.jpg"),
-            thumbnailStyleVersion: WalkRouteThumbnailGenerator.styleVersion
-        )
-
-        #expect(recording.localThumbnailFileExists == false)
-    }
-
-    @Test("Conflict timestamps fall back for existing user-edited recordings")
-    func conflictTimestampFallback() {
-        let createdAt = Date(timeIntervalSince1970: 1_000)
-        let recording = WalkRecording(
-            title: "Edited Walk",
-            createdAt: createdAt,
-            mode: .walk,
-            isTitleUserEdited: true,
-            isDescriptionUserEdited: true
-        )
-
-        #expect(recording.titleConflictTimestamp == createdAt)
-        #expect(recording.descriptionConflictTimestamp == createdAt)
-
-        let editedAt = Date(timeIntervalSince1970: 2_000)
-        recording.titleEditedAt = editedAt
-        recording.descriptionEditedAt = editedAt
-
-        #expect(recording.titleConflictTimestamp == editedAt)
-        #expect(recording.descriptionConflictTimestamp == editedAt)
-    }
-
-    @Test("Delete confirmation explains local fallback video cleanup")
-    func deleteConfirmationForLocalVideoFallback() {
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoURL: URL(fileURLWithPath: "/tmp/video.mov")
-        )
-
-        #expect(recording.deleteConfirmationMessage.contains("app-managed video file"))
-    }
-
-    @Test("Delete confirmation explains local video cleanup without removing Photos copies")
-    func deleteConfirmationForLocalVideoWithPhotosCopy() {
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoURL: URL(fileURLWithPath: "/tmp/video.mov"),
-            videoAssetIdentifier: "photos-asset-id"
-        )
-
-        #expect(recording.deleteConfirmationMessage.contains("app-managed video file"))
-        #expect(recording.deleteConfirmationMessage.contains("Photos copy remains"))
-    }
-
-    @Test("Delete confirmation explains route-only cleanup")
-    func deleteConfirmationForRouteOnlyWalk() {
-        let recording = WalkRecording(title: "Walk", mode: .walk)
-
-        #expect(recording.deleteConfirmationMessage == "This permanently removes the recording and its route.")
-    }
-}
-
-@MainActor
-struct WalkRecordingSessionTests {
-    @Test("The first accurate location starts the route")
-    func acceptsFirstAccurateLocation() {
-        let now = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: now)
-        let location = makeLocation(latitude: 33, longitude: -112, accuracy: 5, timestamp: now)
-
-        #expect(session.accept(location, now: now))
-        #expect(session.snapshot.points.count == 1)
-        #expect(session.snapshot.distanceMeters == 0)
-    }
-
-    @Test("Inaccurate and stale locations are rejected")
-    func rejectsLowQualityLocations() {
-        let now = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: now)
-        let inaccurate = makeLocation(latitude: 33, longitude: -112, accuracy: 75, timestamp: now)
-        let stale = makeLocation(
-            latitude: 33,
-            longitude: -112,
-            accuracy: 5,
-            timestamp: now.addingTimeInterval(-30)
-        )
-
-        #expect(session.accept(inaccurate, now: now) == false)
-        #expect(session.accept(stale, now: now) == false)
-        #expect(session.snapshot.points.isEmpty)
-    }
-
-    @Test("Noise below the movement threshold is rejected")
-    func rejectsNoise() {
-        let now = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: now)
-        let first = makeLocation(latitude: 33, longitude: -112, accuracy: 5, timestamp: now)
-        let nearby = makeLocation(
-            latitude: 33.000001,
-            longitude: -112,
-            accuracy: 5,
-            timestamp: now.addingTimeInterval(2)
-        )
-
-        #expect(session.accept(first, now: now))
-        #expect(session.accept(nearby, now: now.addingTimeInterval(2)) == false)
-        #expect(session.snapshot.points.count == 1)
-    }
-
-    @Test("Accepted movement increases the route distance")
-    func accumulatesDistance() {
-        let now = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: now)
-        let first = makeLocation(latitude: 33, longitude: -112, accuracy: 5, timestamp: now)
-        let second = makeLocation(
-            latitude: 33.001,
-            longitude: -112,
-            accuracy: 5,
-            timestamp: now.addingTimeInterval(10)
-        )
-
-        #expect(session.accept(first, now: now))
-        #expect(session.accept(second, now: now.addingTimeInterval(10)))
-        #expect(session.snapshot.points.count == 2)
-        #expect(session.snapshot.distanceMeters > 100)
-    }
-
-    @Test("Duration is measured from the session start")
-    func updatesDuration() {
-        let start = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: start)
-
-        session.updateDuration(at: start.addingTimeInterval(125))
-
-        #expect(session.snapshot.duration == 125)
-        #expect(session.snapshot.duration.timerText == "2:05")
-    }
-
-    @Test("A video walk session creates video walk metadata")
-    func createsVideoWalkSession() {
-        let start = Date(timeIntervalSince1970: 1_000)
-        let session = WalkRecordingSession(startedAt: start, mode: .videoWalk)
-
-        #expect(session.snapshot.mode == .videoWalk)
-        #expect(session.snapshot.title.hasPrefix("Video Walk"))
-        #expect(session.snapshot.hasVideo == false)
-    }
-
-    private func makeLocation(
-        latitude: CLLocationDegrees,
-        longitude: CLLocationDegrees,
-        accuracy: CLLocationAccuracy,
-        timestamp: Date
-    ) -> CLLocation {
-        CLLocation(
-            coordinate: CLLocationCoordinate2D(latitude: latitude, longitude: longitude),
-            altitude: 300,
-            horizontalAccuracy: accuracy,
-            verticalAccuracy: 5,
-            course: -1,
-            speed: 1.2,
-            timestamp: timestamp
-        )
-    }
-}
-
-@MainActor
-struct RecordingCoordinatorTests {
-    @Test("A GPS walk blocks a second video walk from starting")
-    func gpsWalkBlocksVideoWalk() {
-        let coordinator = RecordingCoordinator(activeMode: .walk)
-
-        #expect(coordinator.hasActiveRecording)
-        #expect(coordinator.canStart(.walk))
-        #expect(coordinator.canStart(.videoWalk) == false)
-        #expect(coordinator.blockingMode(for: .videoWalk) == .walk)
-        #expect(coordinator.activeTab == .walk)
-    }
-
-    @Test("A video walk blocks a second GPS walk from starting")
-    func videoWalkBlocksGPSWalk() {
-        let coordinator = RecordingCoordinator(activeMode: .videoWalk)
-
-        #expect(coordinator.hasActiveRecording)
-        #expect(coordinator.canStart(.videoWalk))
-        #expect(coordinator.canStart(.walk) == false)
-        #expect(coordinator.blockingMode(for: .walk) == .videoWalk)
-        #expect(coordinator.activeTab == .videoWalk)
-    }
-
-    @Test("An idle coordinator allows either recording mode")
-    func idleCoordinatorAllowsAnyMode() {
-        let coordinator = RecordingCoordinator()
-
-        #expect(coordinator.hasActiveRecording == false)
-        #expect(coordinator.canStart(.walk))
-        #expect(coordinator.canStart(.videoWalk))
-        #expect(coordinator.blockingMode(for: .walk) == nil)
-        #expect(coordinator.blockingMode(for: .videoWalk) == nil)
-        #expect(coordinator.activeTab == nil)
-    }
-
-    @Test("Denied location can recover after Settings grants access", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func deniedLocationCanRecoverAfterSettingsGrant() {
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .denied)
-        let recorder = WalkRecorder(locationClient: locationClient)
-
-        recorder.startPreviewingLocation()
-
-        #expect(recorder.isLocationAccessDenied)
-        #expect(recorder.isPreviewingLocation == false)
-        #expect(recorder.errorMessage == "Location access is unavailable.")
-
-        locationClient.authorizationStatus = .authorizedWhenInUse
-        recorder.refreshAuthorizationStatus()
-        recorder.startPreviewingLocation(requestAuthorization: false)
-
-        #expect(recorder.isLocationAccessDenied == false)
-        #expect(recorder.isPreviewingLocation)
-        #expect(recorder.errorMessage == nil)
-        #expect(locationClient.didStartHeadingUpdates)
-    }
-
-    @Test("Background GPS requests Always when only When In Use is granted", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func backgroundGPSRequestsAlwaysUpgrade() {
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedWhenInUse)
-        let recorder = WalkRecorder(locationClient: locationClient)
-
-        recorder.setBackgroundRecordingEnabled(true)
-
-        #expect(locationClient.requestAlwaysAuthorizationCount == 1)
-        #expect(locationClient.requestWhenInUseAuthorizationCount == 0)
-        #expect(recorder.canContinueInBackground == false)
-        #expect(locationClient.allowsBackgroundLocationUpdates == false)
-        #expect(recorder.needsAlwaysLocationForBackgroundRecording)
-    }
-
-    @Test("Background GPS enables updates only for Always-authorized GPS walks", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func backgroundGPSEnablesUpdatesForAlwaysAuthorizedWalks() async throws {
-        let container = try makeTestContainer()
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedAlways)
-        let recorder = WalkRecorder(locationClient: locationClient)
-        let coordinator = RecordingCoordinator(recorder: recorder)
-
-        let didStart = await coordinator.start(
-            in: container.mainContext,
-            mode: .walk,
-            allowsBackgroundRecording: true
-        )
-
-        #expect(didStart)
-        #expect(recorder.canContinueInBackground)
-        #expect(locationClient.allowsBackgroundLocationUpdates)
-        #expect(locationClient.pausesLocationUpdatesAutomatically == false)
-        #expect(locationClient.backgroundActivityCount == 1)
-        let backgroundActivity = try #require(locationClient.backgroundActivities.first)
-
-        await coordinator.discard()
-
-        #expect(coordinator.hasActiveRecording == false)
-        #expect(recorder.phase == .ready)
-        #expect(locationClient.allowsBackgroundLocationUpdates == false)
-        #expect(locationClient.pausesLocationUpdatesAutomatically)
-        #expect(backgroundActivity.didInvalidate)
-        #expect(locationClient.backgroundActivityCount == 1)
-    }
-
-    @Test("Saving an Always-authorized background GPS walk invalidates background activity without replacing it", .bug("https://github.com/heathdj/ASMRWalk/issues/29"))
-    func savingBackgroundGPSWalkInvalidatesBackgroundActivityWithoutReplacement() async throws {
-        let container = try makeTestContainer()
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedAlways)
-        let recorder = WalkRecorder(locationClient: locationClient)
-        let coordinator = RecordingCoordinator(recorder: recorder)
-
-        let didStart = await coordinator.start(
-            in: container.mainContext,
-            mode: .walk,
-            allowsBackgroundRecording: true
-        )
-
-        try #require(didStart)
-        #expect(locationClient.allowsBackgroundLocationUpdates)
-        #expect(locationClient.backgroundActivityCount == 1)
-        let backgroundActivity = try #require(locationClient.backgroundActivities.first)
-
-        await coordinator.stopAndSave()
-
-        #expect(coordinator.hasActiveRecording == false)
-        #expect(recorder.phase == .ready)
-        #expect(locationClient.allowsBackgroundLocationUpdates == false)
-        #expect(locationClient.pausesLocationUpdatesAutomatically)
-        #expect(backgroundActivity.didInvalidate)
-        #expect(locationClient.backgroundActivityCount == 1)
-    }
-
-    @Test("Video walks stay foreground-only even when background GPS is enabled", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func videoWalksStayForegroundOnlyWhenBackgroundGPSIsEnabled() async throws {
-        let container = try makeTestContainer()
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedAlways)
-        let recorder = WalkRecorder(locationClient: locationClient)
-        let coordinator = RecordingCoordinator(recorder: recorder)
-
-        let didStart = await coordinator.start(
-            in: container.mainContext,
-            mode: .videoWalk,
-            allowsBackgroundRecording: true
-        )
-
-        #expect(didStart)
-        #expect(recorder.canContinueInBackground == false)
-        #expect(locationClient.allowsBackgroundLocationUpdates == false)
-        #expect(locationClient.backgroundActivityCount == 0)
-
-        await coordinator.discard()
-    }
-
-    @Test("Video stop saves the local video URL through the coordinator flow", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func videoStopSavesLocalVideoThroughCoordinatorFlow() async throws {
-        let container = try makeTestContainer()
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedWhenInUse)
-        let recorder = WalkRecorder(locationClient: locationClient)
-        let coordinator = RecordingCoordinator(recorder: recorder)
-        let videoURL = URL(fileURLWithPath: "/tmp/local-video.mov")
-        let camera = FakeVideoRecordingController(stopResult: .success(videoURL))
-
-        let didStart = await coordinator.start(in: container.mainContext, mode: .videoWalk)
-        try #require(didStart)
-
-        let result = await VideoWalkStopFlow(
-            coordinator: coordinator,
-            camera: camera
-        ).stop(confirmShortRecording: false)
-
-        #expect(result == .saved)
-        #expect(coordinator.hasActiveRecording == false)
-        #expect(camera.messages.contains("Video walk saved in ASMR Walk."))
-
-        let recording = try #require(try fetchOnlyRecording(in: container))
-        #expect(recording.videoURL == videoURL)
-        #expect(recording.videoAssetIdentifier == nil)
-    }
-
-    @Test("Camera stop failure discards the draft recording through the coordinator flow", .bug("https://github.com/heathdj/ASMRWalk/issues/34"))
-    func cameraStopFailureDiscardsDraftRecordingThroughCoordinatorFlow() async throws {
-        let container = try makeTestContainer()
-        let locationClient = FakeWalkLocationClient(authorizationStatus: .authorizedWhenInUse)
-        let recorder = WalkRecorder(locationClient: locationClient)
-        let coordinator = RecordingCoordinator(recorder: recorder)
-        let camera = FakeVideoRecordingController(stopResult: .failure(TestError(message: "Video stop failed")))
-
-        let didStart = await coordinator.start(in: container.mainContext, mode: .videoWalk)
-        try #require(didStart)
-
-        let result = await VideoWalkStopFlow(
-            coordinator: coordinator,
-            camera: camera
-        ).stop(stopSessionWhenFinished: true, confirmShortRecording: false)
-
-        #expect(result == .failed)
-        #expect(coordinator.hasActiveRecording == false)
-        #expect(recorder.phase == .ready)
-        #expect(camera.messages.contains("Video stop failed"))
-        #expect(camera.didStopSession)
-        #expect(try fetchOnlyRecording(in: container) == nil)
-    }
-}
-
-@MainActor
-struct WalkRecordingTests {
-    @Test("A new walk stores its metadata and defaults")
-    func recordingMetadata() {
-        let createdAt = Date(timeIntervalSince1970: 1_700_000_000)
-        let recording = WalkRecording(
-            title: "Test Walk",
-            createdAt: createdAt,
-            duration: 300,
-            distanceMeters: 800,
-            mode: .walk
-        )
-
-        #expect(recording.title == "Test Walk")
-        #expect(recording.createdAt == createdAt)
-        #expect(recording.duration == 300)
-        #expect(recording.distanceMeters == 800)
-        #expect(recording.mode == .walk)
-        #expect(recording.walkDescription == "")
-        #expect(recording.generatedPlaceName == nil)
-        #expect(recording.metadataGeneratedAt == nil)
-        #expect(recording.isTitleUserEdited == false)
-        #expect(recording.isDescriptionUserEdited == false)
-        #expect(recording.source == .iPhone)
-        #expect(recording.sourceTitle == "iPhone")
-        #expect(recording.captureDeviceName == nil)
-        #expect(recording.routeTimingStart == createdAt)
-        #expect(recording.routeTimingEnd == createdAt.addingTimeInterval(300))
-        #expect(recording.externalVideoReference == nil)
-        #expect(recording.externalVideoStartedAt == nil)
-        #expect(recording.thumbnailURL == nil)
-        #expect(recording.thumbnailStyleVersion == 0)
-        #expect(recording.points.isEmpty)
-        #expect(recording.hasVideo == false)
-    }
-
-    @Test("Adding a point includes it in time-ordered route data")
-    func addingPoints() {
-        let recording = WalkRecording(title: "Test Walk", mode: .walk)
-        let laterPoint = makePoint(timestamp: 200)
-        let earlierPoint = makePoint(timestamp: 100)
-
-        recording.addPoint(laterPoint)
-        recording.addPoint(earlierPoint)
-
-        #expect(recording.points.count == 2)
-        #expect(recording.pointsInTimeOrder.map(\.timestamp) == [
-            earlierPoint.timestamp,
-            laterPoint.timestamp
-        ])
-    }
-
-    @Test("Video presence is derived from its file URL")
-    func videoPresence() {
-        let walk = WalkRecording(title: "Walk", mode: .walk)
-        let videoWalk = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoURL: URL(filePath: "/test/video.mov")
-        )
-
-        #expect(walk.hasVideo == false)
-        #expect(videoWalk.hasVideo)
-    }
-
-    @Test("Local file cleanup includes video and thumbnail assets")
-    func localFileCleanupIncludesVideoAndThumbnailAssets() {
-        let videoURL = URL(filePath: "/tmp/video.mov")
-        let thumbnailURL = URL(filePath: "/tmp/thumbnail.jpg")
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            videoURL: videoURL,
-            thumbnailURL: thumbnailURL
-        )
-
-        #expect(WalkRecordingLocalFiles.removableURLs(for: recording) == [videoURL, thumbnailURL])
-    }
-
-    @Test("Short recordings use a 10 second save confirmation threshold")
-    func shortRecordingThreshold() {
-        let shortWalk = WalkRecording(title: "Short Walk", duration: 9.9, mode: .walk)
-        let tenSecondWalk = WalkRecording(title: "Ten Second Walk", duration: 10, mode: .walk)
-
-        #expect(WalkRecording.shortRecordingThreshold == 10)
-        #expect(shortWalk.isShortRecording)
-        #expect(tenSecondWalk.isShortRecording == false)
-    }
-
-    @Test("Watch recording metadata has safe presentation defaults")
-    func watchRecordingMetadataPresentation() {
-        let startedAt = Date(timeIntervalSince1970: 1_000)
-        let endedAt = Date(timeIntervalSince1970: 1_300)
-        let recording = WalkRecording(
-            title: "Watch Walk",
-            createdAt: startedAt,
-            duration: 300,
-            mode: .walk,
-            recordingSource: .appleWatch,
-            captureDeviceName: "David's Apple Watch",
-            routeStartedAt: startedAt,
-            routeEndedAt: endedAt
-        )
-
-        #expect(recording.source == .appleWatch)
-        #expect(recording.sourceTitle == "Apple Watch")
-        #expect(recording.isWatchRecording)
-        #expect(recording.routeTimingStart == startedAt)
-        #expect(recording.routeTimingEnd == endedAt)
-        #expect(recording.externalCameraTimingMessage == "No external camera timing has been attached.")
-    }
-
-    @Test("External camera timing describes its route offset")
-    func externalCameraTimingPresentation() {
-        let routeStart = Date(timeIntervalSince1970: 1_000)
-        let recording = WalkRecording(
-            title: "External Camera Walk",
-            createdAt: routeStart,
-            duration: 300,
-            mode: .walk,
-            recordingSource: .appleWatch,
-            routeStartedAt: routeStart,
-            externalVideoReference: "A-cam clip 012",
-            externalVideoStartedAt: routeStart.addingTimeInterval(-5)
-        )
-
-        #expect(recording.externalVideoReference == "A-cam clip 012")
-        #expect(recording.externalCameraTimingMessage == "External camera timing starts 0:05 before the route.")
-    }
-
-    @Test("Video playback route progress follows recorded point timing")
-    func videoPlaybackRouteProgress() throws {
-        let recording = WalkRecording(
-            title: "Video Walk",
-            mode: .videoWalk,
-            points: [
-                makePoint(timestamp: 220, latitude: 33.4500, longitude: -112.0710),
-                makePoint(timestamp: 100, latitude: 33.4484, longitude: -112.0740),
-                makePoint(timestamp: 160, latitude: 33.4490, longitude: -112.0728)
-            ]
-        )
-
-        #expect(try #require(recording.playbackPoint(at: 0)).latitude == 33.4484)
-        #expect(try #require(recording.playbackPoint(at: 65)).latitude == 33.4490)
-        #expect(try #require(recording.playbackPoint(at: 500)).latitude == 33.4500)
-    }
-
-    @Test("Duration presentation handles minutes and hours")
-    func durationPresentation() {
-        let shortWalk = WalkRecording(title: "Short", duration: 125, mode: .walk)
-        let longWalk = WalkRecording(title: "Long", duration: 3_725, mode: .walk)
-
-        #expect(shortWalk.durationText == "2:05")
-        #expect(longWalk.durationText == "1:02:05")
-    }
-
-    @Test("Location points expose their map coordinate")
-    func locationCoordinate() {
-        let point = makePoint(timestamp: 100)
-
-        #expect(point.coordinate.latitude == point.latitude)
-        #expect(point.coordinate.longitude == point.longitude)
-    }
-
-    @Test("Generated recording metadata uses place name for title")
-    func generatedRecordingMetadata() throws {
-        let metadata = try #require(WalkRecordingMetadataBuilder.metadata(
-            for: WalkPlaceMetadata(
-                name: "Papago Park",
-                subLocality: "Camelback East",
-                locality: "Phoenix",
-                administrativeArea: "Arizona",
-                country: "United States"
-            ),
-            mode: .videoWalk,
-            createdAt: Date(timeIntervalSince1970: 1_700_000_000),
-            duration: 125,
-            distanceMeters: 840,
-            generatedAt: Date(timeIntervalSince1970: 2_000)
-        ))
-
-        #expect(metadata.title == "Papago Park Video Walk")
-        #expect(metadata.placeName == "Papago Park")
-        #expect(metadata.walkDescription.contains("Video Walk near Papago Park"))
-        #expect(metadata.walkDescription.contains("2:05"))
-        #expect(metadata.generatedAt == Date(timeIntervalSince1970: 2_000))
-    }
-
-    @Test("Generated recording metadata falls back when no place name exists")
-    func generatedRecordingMetadataRequiresPlaceName() {
-        let metadata = WalkRecordingMetadataBuilder.metadata(
-            for: WalkPlaceMetadata(),
-            mode: .walk,
-            createdAt: Date(timeIntervalSince1970: 1_700_000_000),
-            duration: 125,
-            distanceMeters: 840
-        )
-
-        #expect(metadata == nil)
-    }
-
-    @Test("Route metadata uses the middle point for place lookup")
-    func routeMetadataRepresentativeCoordinate() throws {
-        let createdAt = Date(timeIntervalSince1970: 1_000)
-        let snapshot = makeSnapshot(
-            id: UUID(),
-            createdAt: createdAt,
-            duration: 30,
-            distanceMeters: 90,
-            pointCount: 5
-        )
-
-        let coordinate = try #require(WalkRecordingMetadataGenerator.representativeCoordinate(from: snapshot.points))
-
-        #expect(coordinate.latitude == snapshot.points[2].latitude)
-        #expect(coordinate.longitude == snapshot.points[2].longitude)
-    }
-
-    @Test("Metadata generator updates saved recordings when place lookup succeeds")
-    func metadataGeneratorUpdatesSavedRecordings() async throws {
-        let recordingID = try #require(UUID(uuidString: "45C81AB1-0404-4C6A-AAB2-14AD16437F32"))
-        let generatedAt = Date(timeIntervalSince1970: 4_000)
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let snapshot = makeSnapshot(
-            id: recordingID,
-            title: "Aug 13, 2026 Walk",
-            createdAt: Date(timeIntervalSince1970: 1_800),
-            duration: 95,
-            distanceMeters: 420,
-            pointCount: 5,
-            latitude: 33.5000,
-            longitude: -112.1000
-        )
-
-        try await persistence.save(snapshot)
-
-        await WalkRecordingMetadataGenerator.generate(
-            for: snapshot,
-            in: container,
-            geocoder: FakeWalkPlaceGeocoder(metadata: WalkPlaceMetadata(name: "Encanto Park")),
-            date: generatedAt
-        )
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.title == "Encanto Park Walk")
-        #expect(recording.walkDescription.contains("Walk near Encanto Park"))
-        #expect(recording.generatedPlaceName == "Encanto Park")
-        #expect(recording.metadataGeneratedAt == generatedAt)
-    }
-
-    @Test("Metadata generator leaves saved recordings alone when place lookup fails")
-    func metadataGeneratorIgnoresLookupFailures() async throws {
-        let recordingID = try #require(UUID(uuidString: "075250E1-CF1B-49E6-AC6A-317A7A046470"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let snapshot = makeSnapshot(
-            id: recordingID,
-            title: "Fallback Walk",
-            duration: 95,
-            distanceMeters: 420,
-            pointCount: 5,
-            latitude: 33.6000,
-            longitude: -112.2000
-        )
-
-        try await persistence.save(snapshot)
-
-        await WalkRecordingMetadataGenerator.generate(
-            for: snapshot,
-            in: container,
-            geocoder: FailingWalkPlaceGeocoder()
-        )
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.title == "Fallback Walk")
-        #expect(recording.walkDescription.isEmpty)
-        #expect(recording.generatedPlaceName == nil)
-        #expect(recording.metadataGeneratedAt == nil)
-    }
-
-    @Test("Recordings can be inserted, fetched, updated, and deleted")
-    func recordingLifecycle() throws {
-        let container = try makeContainer()
-        let context = container.mainContext
-        let recording = WalkRecording(
-            title: "Original Title",
-            duration: 120,
-            distanceMeters: 400,
-            mode: .walk,
-            points: [makePoint(timestamp: 100)]
-        )
-
-        context.insert(recording)
-        try context.save()
-
-        var fetched = try context.fetch(FetchDescriptor<WalkRecording>())
-        let savedRecording = try #require(fetched.first)
-        #expect(fetched.count == 1)
-        #expect(savedRecording.points.count == 1)
-
-        savedRecording.title = "Updated Title"
-        try context.save()
-
-        fetched = try context.fetch(FetchDescriptor<WalkRecording>())
-        #expect(fetched.first?.title == "Updated Title")
-
-        context.delete(savedRecording)
-        try context.save()
-
-        #expect(try context.fetchCount(FetchDescriptor<WalkRecording>()) == 0)
-        #expect(try context.fetchCount(FetchDescriptor<LocationPoint>()) == 0)
-    }
-
-    @Test("Persistence checkpoints append only new route points")
-    func persistenceCheckpointsAppendOnlyNewPoints() async throws {
-        let recordingID = try #require(UUID(uuidString: "E94C5F79-B21C-4687-9345-D640978534E1"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let firstCheckpoint = makeSnapshot(
-            id: recordingID,
-            duration: 30,
-            distanceMeters: 90,
-            pointCount: 10
-        )
-        let secondCheckpoint = makeSnapshot(
-            id: recordingID,
-            duration: 60,
-            distanceMeters: 180,
-            pointCount: 15
-        )
-
-        try await persistence.save(firstCheckpoint)
-        try await persistence.save(secondCheckpoint)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.points.count == 15)
-        #expect(recording.duration == 60)
-        #expect(recording.distanceMeters == 180)
-        #expect(recording.recordingSource == WalkRecordingSource.iPhone.rawValue)
-        #expect(recording.routeStartedAt == firstCheckpoint.routeStartedAt)
-        #expect(recording.routeEndedAt == secondCheckpoint.routeEndedAt)
-        #expect(recording.pointsInTimeOrder.map(\.timestamp) == secondCheckpoint.points.map(\.timestamp))
-        #expect(try ModelContext(container).fetchCount(FetchDescriptor<WalkRecording>()) == 1)
-    }
-
-    @Test("Repeated persistence checkpoints do not duplicate points")
-    func repeatedPersistenceCheckpointsDoNotDuplicatePoints() async throws {
-        let recordingID = try #require(UUID(uuidString: "5E19B546-9E52-4C93-9086-4C263FB8384D"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let checkpoint = makeSnapshot(
-            id: recordingID,
-            duration: 45,
-            distanceMeters: 135,
-            pointCount: 12
-        )
-
-        try await persistence.save(checkpoint)
-        try await persistence.save(checkpoint)
-        try await persistence.save(checkpoint)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.points.count == 12)
-        #expect(Set(recording.points.map(\.timestamp)).count == 12)
-        #expect(try ModelContext(container).fetchCount(FetchDescriptor<WalkRecording>()) == 1)
-    }
-
-    @Test("Schema hardening retains route data and local file references")
-    func schemaHardeningRetainsLocalRecordingData() throws {
-        let container = try makeContainer()
-        let context = container.mainContext
-        let videoURL = URL(filePath: "/tmp/local-video.mov")
-        let recording = WalkRecording(
-            title: "Beta Walk",
-            duration: 180,
-            distanceMeters: 540,
-            mode: .videoWalk,
-            videoURL: videoURL,
-            points: [
-                makePoint(timestamp: 100),
-                makePoint(timestamp: 120)
-            ]
-        )
-
-        context.insert(recording)
-        try context.save()
-
-        let fetched = try #require(try fetchRecording(id: recording.id, in: container))
-        #expect(fetched.title == "Beta Walk")
-        #expect(fetched.videoURL == videoURL)
-        #expect(fetched.points.count == 2)
-        #expect(fetched.pointsInTimeOrder.map(\.timestamp) == recording.pointsInTimeOrder.map(\.timestamp))
-    }
-
-    @Test("Final persistence save retains all points after partial checkpoints")
-    func finalPersistenceSaveRetainsAllPointsAfterPartialCheckpoints() async throws {
-        let recordingID = try #require(UUID(uuidString: "8F23B3D8-6911-4193-A558-D52704386507"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let firstCheckpoint = makeSnapshot(
-            id: recordingID,
-            duration: 120,
-            distanceMeters: 360,
-            pointCount: 40
-        )
-        let finalSnapshot = makeSnapshot(
-            id: recordingID,
-            title: "Finished Walk",
-            duration: 180,
-            distanceMeters: 540,
-            pointCount: 75
-        )
-
-        try await persistence.save(firstCheckpoint)
-        try await persistence.save(finalSnapshot)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.title == "Finished Walk")
-        #expect(recording.points.count == 75)
-        #expect(recording.pointsInTimeOrder.map(\.timestamp) == finalSnapshot.points.map(\.timestamp))
-    }
-
-    @Test("Final persistence save recovers points after a missed checkpoint")
-    func finalPersistenceSaveRecoversPointsAfterMissedCheckpoint() async throws {
-        let recordingID = try #require(UUID(uuidString: "9EB7066D-6430-4F2D-9639-A91E0505D294"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let firstCheckpoint = makeSnapshot(
-            id: recordingID,
-            duration: 120,
-            distanceMeters: 360,
-            pointCount: 40
-        )
-        _ = makeSnapshot(
-            id: recordingID,
-            duration: 150,
-            distanceMeters: 450,
-            pointCount: 60
-        )
-        let finalSnapshot = makeSnapshot(
-            id: recordingID,
-            duration: 210,
-            distanceMeters: 630,
-            pointCount: 90
-        )
-
-        try await persistence.save(firstCheckpoint)
-        try await persistence.save(finalSnapshot)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.points.count == 90)
-        #expect(recording.pointsInTimeOrder.map(\.timestamp) == finalSnapshot.points.map(\.timestamp))
-        #expect(recording.duration == 210)
-        #expect(recording.distanceMeters == 630)
-    }
-
-    @Test("Persistence updates metadata without rewriting existing points")
-    func persistenceUpdatesMetadataWithoutRewritingExistingPoints() async throws {
-        let recordingID = try #require(UUID(uuidString: "2823E0B3-0866-43EA-A1D9-9404C28A0201"))
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let firstCheckpoint = makeSnapshot(
-            id: recordingID,
-            duration: 30,
-            distanceMeters: 90,
-            pointCount: 8
-        )
-        let metadataOnlyCheckpoint = makeSnapshot(
-            id: recordingID,
-            title: "Updated Metadata",
-            duration: 75,
-            distanceMeters: 90,
-            pointCount: 8
-        )
-
-        try await persistence.save(firstCheckpoint)
-        try await persistence.save(metadataOnlyCheckpoint)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.title == "Updated Metadata")
-        #expect(recording.duration == 75)
-        #expect(recording.points.count == 8)
-        #expect(recording.pointsInTimeOrder.map(\.timestamp) == metadataOnlyCheckpoint.points.map(\.timestamp))
-    }
-
-    @Test("Persistence stores Watch source and external camera timing metadata")
-    func persistenceStoresWatchSourceAndExternalCameraTiming() async throws {
-        let recordingID = try #require(UUID(uuidString: "F30E2C38-8714-4FA5-B30D-817AE52A0E77"))
-        let routeStartedAt = Date(timeIntervalSince1970: 2_000)
-        let routeEndedAt = Date(timeIntervalSince1970: 2_420)
-        let externalVideoStartedAt = Date(timeIntervalSince1970: 1_995)
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let snapshot = makeSnapshot(
-            id: recordingID,
-            createdAt: routeStartedAt,
-            duration: 420,
-            distanceMeters: 1_200,
-            pointCount: 12,
-            recordingSource: .appleWatch,
-            captureDeviceName: "Apple Watch Ultra",
-            routeStartedAt: routeStartedAt,
-            routeEndedAt: routeEndedAt,
-            externalVideoReference: "Camera A clip 001",
-            externalVideoStartedAt: externalVideoStartedAt
-        )
-
-        try await persistence.save(snapshot)
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.source == .appleWatch)
-        #expect(recording.captureDeviceName == "Apple Watch Ultra")
-        #expect(recording.routeStartedAt == routeStartedAt)
-        #expect(recording.routeEndedAt == routeEndedAt)
-        #expect(recording.externalVideoReference == "Camera A clip 001")
-        #expect(recording.externalVideoStartedAt == externalVideoStartedAt)
-    }
-
-    @Test("Persistence applies generated metadata without overwriting user edits")
-    func persistenceAppliesGeneratedMetadataWithoutOverwritingUserEdits() async throws {
-        let recordingID = try #require(UUID(uuidString: "EF7D1418-0606-4013-9503-A4B3F609D17A"))
-        let generatedAt = Date(timeIntervalSince1970: 2_000)
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let snapshot = makeSnapshot(
-            id: recordingID,
-            title: "Default Walk",
-            duration: 75,
-            distanceMeters: 210,
-            pointCount: 8
-        )
-
-        try await persistence.save(snapshot)
-        let context = ModelContext(container)
-        let descriptor = FetchDescriptor<WalkRecording>(
-            predicate: #Predicate { recording in
-                recording.id == recordingID
-            }
-        )
-        let recording = try #require(try context.fetch(descriptor).first)
-        recording.title = "User Title"
-        recording.isTitleUserEdited = true
-        try context.save()
-
-        try await persistence.updateGeneratedMetadata(
-            recordingID: recordingID,
-            metadata: WalkGeneratedRecordingMetadata(
-                title: "Generated Title",
-                walkDescription: "Generated description.",
-                placeName: "Papago Park",
-                generatedAt: generatedAt
-            )
-        )
-
-        let updatedRecording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(updatedRecording.title == "User Title")
-        #expect(updatedRecording.walkDescription == "Generated description.")
-        #expect(updatedRecording.generatedPlaceName == "Papago Park")
-        #expect(updatedRecording.metadataGeneratedAt == generatedAt)
-    }
-
-    @Test("Persistence stores generated route thumbnail URLs")
-    func persistenceStoresGeneratedRouteThumbnailURLs() async throws {
-        let recordingID = try #require(UUID(uuidString: "CC9044A6-89AF-4BA9-9B8F-C4CC25E4313F"))
-        let thumbnailURL = URL(filePath: "/tmp/route-thumbnail.jpg")
-        let container = try makeContainer()
-        let persistence = WalkRecordingPersistence(modelContainer: container)
-        let snapshot = makeSnapshot(
-            id: recordingID,
-            title: "Thumbnail Walk",
-            duration: 75,
-            distanceMeters: 210,
-            pointCount: 8
-        )
-
-        try await persistence.save(snapshot)
-        try await persistence.updateThumbnailURL(
-            recordingID: recordingID,
-            thumbnailURL: thumbnailURL,
-            styleVersion: WalkRouteThumbnailGenerator.styleVersion
-        )
-
-        let recording = try #require(try fetchRecording(id: recordingID, in: container))
-        #expect(recording.thumbnailURL == thumbnailURL)
-        #expect(recording.thumbnailStyleVersion == WalkRouteThumbnailGenerator.styleVersion)
-    }
-
-    @Test("Route thumbnail filenames are tied to recording IDs")
-    func routeThumbnailFilenamesAreTiedToRecordingIDs() throws {
-        let recordingID = try #require(UUID(uuidString: "C88807DA-8957-4B6F-A6C7-1D1C37BA9515"))
-        let url = try WalkRouteThumbnailGenerator.thumbnailURL(recordingID: recordingID)
-
-        #expect(WalkRouteThumbnailGenerator.styleVersion == 1)
-        #expect(url.lastPathComponent == "\(recordingID.uuidString).jpg")
-        #expect(url.deletingLastPathComponent().lastPathComponent == "Route Thumbnails")
-    }
-
-    @Test("Sample data contains both recording modes")
-    func sampleData() {
-        #expect(SampleData.recordings.count == 2)
-        #expect(Set(SampleData.recordings.map(\.mode)) == Set(RecordingMode.allCases))
-        #expect(SampleData.recordings.allSatisfy { $0.points.isEmpty == false })
-    }
-
-    @Test("Google Maps export includes the route endpoints and walking mode")
-    func googleMapsExport() throws {
-        let recording = WalkRecording(
-            title: "Export Walk",
-            mode: .walk,
-            points: [
-                makePoint(timestamp: 100, latitude: 33.4484, longitude: -112.0740),
-                makePoint(timestamp: 200, latitude: 33.4490, longitude: -112.0728),
-                makePoint(timestamp: 300, latitude: 33.4500, longitude: -112.0710)
-            ]
-        )
-
-        let url = try #require(WalkRouteExport(recording: recording).googleMapsURL)
-        let components = try #require(URLComponents(url: url, resolvingAgainstBaseURL: false))
-        let query = Dictionary(uniqueKeysWithValues: (components.queryItems ?? []).compactMap { item in
-            item.value.map { (item.name, $0) }
-        })
-
-        #expect(query["origin"] == "33.448400,-112.074000")
-        #expect(query["destination"] == "33.450000,-112.071000")
-        #expect(query["travelmode"] == "walking")
-        #expect(query["waypoints"] == "33.449000,-112.072800")
-    }
-
-    @Test("GPX export preserves every point in time order and escapes its title")
-    func gpxExport() throws {
-        let recording = WalkRecording(
-            title: "Creek & Canal",
-            createdAt: Date(timeIntervalSince1970: 300),
-            mode: .walk,
-            points: [
-                makePoint(timestamp: 200, latitude: 33.4490, longitude: -112.0728),
-                makePoint(timestamp: 100, latitude: 33.4484, longitude: -112.0740)
-            ]
-        )
-
-        let export = WalkRouteExport(recording: recording)
-        let firstPoint = try #require(export.gpxText.range(of: "lat=\"33.448400\""))
-        let secondPoint = try #require(export.gpxText.range(of: "lat=\"33.449000\""))
-
-        #expect(export.gpxText.contains("<name>Creek &amp; Canal</name>"))
-        #expect(export.gpxText.components(separatedBy: "<trkpt ").count - 1 == 2)
-        #expect(firstPoint.lowerBound < secondPoint.lowerBound)
-        #expect(export.gpxFile.filename == "Creek-Canal.gpx")
-    }
-
-    @Test("GPX export includes ASMR Walk extensions for importer sync and diagnostics")
-    func gpxExportExtensions() throws {
-        let recordingID = try #require(UUID(uuidString: "3B278DC8-F7C6-4A01-8B55-C627DD6F00E1"))
-        let recording = WalkRecording(
-            id: recordingID,
-            title: "Video Export",
-            createdAt: Date(timeIntervalSince1970: 300),
-            duration: 142.75,
-            mode: .videoWalk,
-            videoURL: URL(filePath: "/private/var/mobile/Containers/Data/Application/video.mov"),
-            points: [
-                makePoint(timestamp: 100, horizontalAccuracy: 4.25, speed: 1.5),
-                makePoint(timestamp: 120, latitude: 33.4490, longitude: -112.0728, horizontalAccuracy: 8)
-            ]
-        )
-
-        let gpxText = WalkRouteExport(recording: recording).gpxText
-
-        #expect(gpxText.contains("xmlns:asmrwalk=\"https://asmrwalk.app/gpx/1\""))
-        #expect(gpxText.contains("<asmrwalk:recordingID>\(recordingID.uuidString)</asmrwalk:recordingID>"))
-        #expect(gpxText.contains("<asmrwalk:durationSeconds>142.75</asmrwalk:durationSeconds>"))
-        #expect(gpxText.contains("<asmrwalk:recordingMode>videoWalk</asmrwalk:recordingMode>"))
-        #expect(gpxText.contains("<asmrwalk:recordingSource>iPhone</asmrwalk:recordingSource>"))
-        #expect(gpxText.contains("<asmrwalk:routeStartedAt>1970-01-01T00:05:00Z</asmrwalk:routeStartedAt>"))
-        #expect(gpxText.contains("<asmrwalk:routeEndedAt>1970-01-01T00:07:22Z</asmrwalk:routeEndedAt>"))
-        #expect(gpxText.contains("<asmrwalk:hasVideo>true</asmrwalk:hasVideo>"))
-        #expect(gpxText.contains("<asmrwalk:horizontalAccuracyMeters>4.25</asmrwalk:horizontalAccuracyMeters>"))
-        #expect(gpxText.contains("<asmrwalk:horizontalAccuracyMeters>8</asmrwalk:horizontalAccuracyMeters>"))
-        #expect(gpxText.contains("<asmrwalk:speedMetersPerSecond>1.5</asmrwalk:speedMetersPerSecond>"))
-        #expect(gpxText.components(separatedBy: "<asmrwalk:speedMetersPerSecond>").count - 1 == 1)
-        #expect(gpxText.contains("video.mov") == false)
-        #expect(gpxText.contains("/private/var/mobile") == false)
-    }
-
-    @Test("GPX export includes Watch and external camera timing metadata")
-    func gpxExportIncludesWatchAndExternalCameraMetadata() {
-        let routeStartedAt = Date(timeIntervalSince1970: 1_000)
-        let routeEndedAt = Date(timeIntervalSince1970: 1_240)
-        let externalVideoStartedAt = Date(timeIntervalSince1970: 995)
-        let recording = WalkRecording(
-            title: "Watch Export",
-            createdAt: routeStartedAt,
-            duration: 240,
-            mode: .walk,
-            recordingSource: .appleWatch,
-            captureDeviceName: "David's Apple Watch & Camera",
-            routeStartedAt: routeStartedAt,
-            routeEndedAt: routeEndedAt,
-            externalVideoReference: "A-cam <clip 12>",
-            externalVideoStartedAt: externalVideoStartedAt,
-            points: [
-                makePoint(timestamp: 1_000)
-            ]
-        )
-
-        let gpxText = WalkRouteExport(recording: recording).gpxText
-
-        #expect(gpxText.contains("<asmrwalk:recordingSource>appleWatch</asmrwalk:recordingSource>"))
-        #expect(gpxText.contains("<asmrwalk:captureDeviceName>David&apos;s Apple Watch &amp; Camera</asmrwalk:captureDeviceName>"))
-        #expect(gpxText.contains("<asmrwalk:routeStartedAt>1970-01-01T00:16:40Z</asmrwalk:routeStartedAt>"))
-        #expect(gpxText.contains("<asmrwalk:routeEndedAt>1970-01-01T00:20:40Z</asmrwalk:routeEndedAt>"))
-        #expect(gpxText.contains("<asmrwalk:externalVideoReference>A-cam &lt;clip 12&gt;</asmrwalk:externalVideoReference>"))
-        #expect(gpxText.contains("<asmrwalk:externalVideoStartedAt>1970-01-01T00:16:35Z</asmrwalk:externalVideoStartedAt>"))
-    }
-
-    @Test("GPX export includes recording descriptions")
-    func gpxExportIncludesDescription() {
-        let recording = WalkRecording(
-            title: "Described Walk",
-            mode: .walk,
-            walkDescription: "A quiet canal walk near Phoenix & Tempe.",
-            points: [
-                makePoint(timestamp: 100)
-            ]
-        )
-
-        let gpxText = WalkRouteExport(recording: recording).gpxText
-
-        #expect(gpxText.contains("<desc>A quiet canal walk near Phoenix &amp; Tempe.</desc>"))
-        #expect(gpxText.contains("<asmrwalk:description>A quiet canal walk near Phoenix &amp; Tempe.</asmrwalk:description>"))
-    }
-
-    @Test("GPX export uses POSIX-safe number formatting for elevation")
-    func gpxExportElevationFormatting() {
-        let recording = WalkRecording(
-            title: "Elevation Export",
-            mode: .walk,
-            points: [
-                makePoint(timestamp: 100, altitude: 331.25)
-            ]
-        )
-
-        #expect(WalkRouteExport(recording: recording).gpxText.contains("<ele>331.25</ele>"))
-    }
-
-    private func makeContainer() throws -> ModelContainer {
-        let configuration = ModelConfiguration(isStoredInMemoryOnly: true)
-        return try ModelContainer(
-            for: WalkRecording.self,
-            LocationPoint.self,
-            configurations: configuration
-        )
-    }
-
-    private func makePoint(
-        timestamp: TimeInterval,
-        latitude: Double = 33.4484,
-        longitude: Double = -112.0740,
-        altitude: Double? = nil,
-        horizontalAccuracy: Double = 5,
-        speed: Double? = nil
-    ) -> LocationPoint {
-        LocationPoint(
-            timestamp: Date(timeIntervalSince1970: timestamp),
-            latitude: latitude,
-            longitude: longitude,
-            altitude: altitude,
-            horizontalAccuracy: horizontalAccuracy,
-            speed: speed
-        )
-    }
-
-    private func makeSnapshot(
-        id: UUID,
-        title: String = "Checkpoint Walk",
-        createdAt: Date = Date(timeIntervalSince1970: 1_000),
-        duration: TimeInterval,
-        distanceMeters: Double,
-        pointCount: Int,
-        latitude: Double = 33.4484,
-        longitude: Double = -112.0740,
-        recordingSource: WalkRecordingSource = .iPhone,
-        captureDeviceName: String? = nil,
-        routeStartedAt: Date? = nil,
-        routeEndedAt: Date? = nil,
-        externalVideoReference: String? = nil,
-        externalVideoStartedAt: Date? = nil
-    ) -> WalkRecordingSnapshot {
-        WalkRecordingSnapshot(
-            id: id,
-            title: title,
-            createdAt: createdAt,
-            duration: duration,
-            distanceMeters: distanceMeters,
-            mode: .walk,
-            recordingSource: recordingSource,
-            captureDeviceName: captureDeviceName,
-            routeStartedAt: routeStartedAt ?? createdAt,
-            routeEndedAt: routeEndedAt ?? createdAt.addingTimeInterval(duration),
-            externalVideoReference: externalVideoReference,
-            externalVideoStartedAt: externalVideoStartedAt,
-            points: (0..<pointCount).map { index in
-                LocationPointSnapshot(
-                    timestamp: createdAt.addingTimeInterval(TimeInterval(index)),
-                    latitude: latitude + (Double(index) * 0.0001),
-                    longitude: longitude + (Double(index) * 0.0001),
-                    altitude: nil,
-                    horizontalAccuracy: 5,
-                    speed: 1.2
-                )
-            }
-        )
-    }
-
-    private func fetchRecording(id: UUID, in container: ModelContainer) throws -> WalkRecording? {
-        let context = ModelContext(container)
-        let descriptor = FetchDescriptor<WalkRecording>(
-            predicate: #Predicate { recording in
-                recording.id == id
-            }
-        )
-        return try context.fetch(descriptor).first
-    }
-}
-
-private func makeTestContainer() throws -> ModelContainer {
-    let configuration = ModelConfiguration(isStoredInMemoryOnly: true)
-    return try ModelContainer(
-        for: WalkRecording.self,
-        LocationPoint.self,
-        configurations: configuration
-    )
-}
-
-private func fetchOnlyRecording(in container: ModelContainer) throws -> WalkRecording? {
-    let context = ModelContext(container)
-    let recordings = try context.fetch(FetchDescriptor<WalkRecording>())
-    return recordings.first
-}
-
-nonisolated private struct TestError: LocalizedError {
-    let message: String
-
-    var errorDescription: String? {
-        message
-    }
-}
-
-nonisolated private struct FakeWalkPlaceGeocoder: WalkPlaceGeocoding {
-    let metadata: WalkPlaceMetadata
-
-    func place(for coordinate: WalkRouteMetadataCoordinate) async throws -> WalkPlaceMetadata {
-        metadata
-    }
-}
-
-nonisolated private struct FailingWalkPlaceGeocoder: WalkPlaceGeocoding {
-    func place(for coordinate: WalkRouteMetadataCoordinate) async throws -> WalkPlaceMetadata {
-        throw TestError(message: "Place lookup failed")
-    }
-}
-
-@MainActor
-private final class FakeBackgroundActivity: WalkBackgroundActivity {
-    private(set) var didInvalidate = false
-
-    func invalidate() {
-        didInvalidate = true
-    }
-}
-
-@MainActor
-private final class FakeWalkLocationClient: WalkLocationClient {
-    var authorizationStatus: CLAuthorizationStatus
-    var allowsBackgroundLocationUpdates = false
-    var pausesLocationUpdatesAutomatically = true
-    var headingUpdatesAvailable = true
-    private(set) var requestWhenInUseAuthorizationCount = 0
-    private(set) var requestAlwaysAuthorizationCount = 0
-    private(set) var didStartHeadingUpdates = false
-    private(set) var didStopHeadingUpdates = false
-    private(set) var backgroundActivityCount = 0
-    private(set) var backgroundActivities: [FakeBackgroundActivity] = []
-    private weak var delegate: CLLocationManagerDelegate?
-
-    init(authorizationStatus: CLAuthorizationStatus) {
-        self.authorizationStatus = authorizationStatus
-    }
-
-    func setDelegate(_ delegate: CLLocationManagerDelegate?) {
-        self.delegate = delegate
-    }
-
-    func requestWhenInUseAuthorization() {
-        requestWhenInUseAuthorizationCount += 1
-    }
-
-    func requestAlwaysAuthorization() {
-        requestAlwaysAuthorizationCount += 1
-    }
-
-    func startUpdatingHeading() {
-        didStartHeadingUpdates = true
-    }
-
-    func stopUpdatingHeading() {
-        didStopHeadingUpdates = true
-    }
-
-    func liveUpdates() -> AsyncThrowingStream<WalkLocationUpdateSnapshot, Error> {
-        AsyncThrowingStream { continuation in
-            continuation.finish()
-        }
-    }
-
-    func makeBackgroundActivity() -> any WalkBackgroundActivity {
-        backgroundActivityCount += 1
-        let activity = FakeBackgroundActivity()
-        backgroundActivities.append(activity)
-        return activity
-    }
-}
-
-@MainActor
-private final class FakeVideoRecordingController: VideoRecordingControlling {
-    let stopURL: URL?
-    let stopError: (any Error)?
-    private(set) var messages: [String] = []
-    private(set) var errors: [String] = []
-    private(set) var didStopSession = false
-
-    init(stopResult: Result<URL, TestError>) {
-        switch stopResult {
-        case let .success(url):
-            stopURL = url
-            stopError = nil
-        case let .failure(error):
-            stopURL = nil
-            stopError = error
-        }
-    }
-
-    func stopRecording() async throws -> URL {
-        if let stopURL {
-            return stopURL
-        }
-
-        throw stopError ?? TestError(message: "Video stop failed")
-    }
-
-    func report(_ error: Error) {
-        errors.append(error.localizedDescription)
-    }
-
-    func reportMessage(_ message: String) {
-        messages.append(message)
-    }
-
-    func stopSession() {
-        didStopSession = true
     }
 }
